Add poll viewing and submitting functionality.
On page load, the app should render a list of poll questions.
This commit adds the QuestionsList page.
Upon clicking on a specific question from QuestionsList,
the user will see a specific question detail, which is represented by
the QuestionsDetail component.
The QuestionsDetail component utilizes material-ui third-party components,
which saved development time in creating a modal with vote-taking
capabilities.
The user is able to vote on a specific poll once inside the QuestionsDetail page.
After a vote is saved, the user lands back inside the QuestionsList page.